Back-end/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.3
django-extensions==3.2.3
PyJWT[crypto]==2.8.0
cryptography==42.0.5
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
